Feature/importer

+++ internal/lambdas/data-import/main.go
@@ -44,16 +44,19 @@ func HandleRequest(ctx context.Context, event awsutil.Event) (string, error) {
 		return "", err
 	}
 
-	// Print contents of /tmp directory... AWS reuses nodes, we may have left files in the past
-	localFS := fileaccess.FSAccess{}
-	tmpFiles, err := localFS.ListObjects("/tmp", "")
-	if err == nil {
-		for c, tmpFile := range tmpFiles {
-			fmt.Printf("%v: %v\n", c+1, tmpFile)
+	// Was used to try to diagnose issues, but not needed as it appears /tmp is empty when we start!
+	/*
+		// Print contents of /tmp directory... AWS reuses nodes, we may have left files in the past
+		localFS := fileaccess.FSAccess{}
+		tmpFiles, err := localFS.ListObjects("/tmp", "")
+		if err == nil {
+			for c, tmpFile := range tmpFiles {
+				fmt.Printf("%v: %v\n", c+1, tmpFile)
+			}
+		} else {
+			fmt.Printf("Failed to list tmp files: %v\n", err)
 		}
-	} else {
-		fmt.Printf("Failed to list tmp files: %v\n", err)
-	}
+	*/
 
 	remoteFS := fileaccess.MakeS3Access(svc)
 
@@ -68,11 +71,13 @@ func HandleRequest(ctx context.Context, event awsutil.Event) (string, error) {
 
 		// Delete the working directory here, there's no point leaving it on a lambda machine, we can't debug it
 		// but if this code ran elsewhere we wouldn't delete it, to have something to look at
-		removeErr := os.RemoveAll(workingDir)
-		if removeErr == nil {
-			fmt.Printf("Failed to remove working dir: \"%v\". Error: %v\n", workingDir, removeErr)
-		} else {
-			fmt.Printf("Removed working dir: \"%v\"\n", workingDir)
+		if len(workingDir) > 0 {
+			removeErr := os.RemoveAll(workingDir)
+			if removeErr == nil {
+				fmt.Printf("Failed to remove working dir: \"%v\". Error: %v\n", workingDir, removeErr)
+			} else {
+				fmt.Printf("Removed working dir: \"%v\"\n", workingDir)
+			}
 		}
 
 		if err != nil {
